refactor for aggregation methods
