+++ src/api/route/movie.route.ts
@@ -2,13 +2,13 @@ import * as express from "express";
 import {
   moviesByGenresAndYear,
   searchMovies,
-  searchMoviesAggregate,
+  searchMoviesYearRange,
 } from "../controller/movie.controller";
 
 const movieRouter = express.Router();
 
 movieRouter.get("/", searchMovies);
 movieRouter.get("/by-genre/:year/:genres", moviesByGenresAndYear);
-movieRouter.get("/agg-search", searchMoviesAggregate);
+movieRouter.get("/agg-search", searchMoviesYearRange);
 
 export { movieRouter };
